Correction de la fonctionnalité de suppression en batch et ajout de la comparaison "est différent de" dans le filtre "string"

+++ src/Model/ModelManager.php
@@ -221,7 +221,13 @@ class ModelManager implements ModelManagerInterface
         }
 
         try {
-            $this->getClassRepository($class)->deleteByIds($queryProxy->getIdentifiers());
+            if ($queryProxy->hasIdentifiers()) {
+                $this->getClassRepository($class)->deleteByIds($queryProxy->getIdentifiers());
+
+                return;
+            }
+
+            $this->getClassRepository($class)->deleteBy($queryProxy->getCriteria());
         } catch (\Exception $e) {
             throw new ModelManagerException(
                 sprintf('Failed to delete object: %s', $class),
